+++ packages/compartment-mapper/src/parse-json.js
@@ -20,6 +20,7 @@ const textDecoder = new TextDecoder();
 /** @type {ParseFn} */
 export const parseJson = (bytes, _specifier, location, _packageLocation) => {
   const source = textDecoder.decode(bytes);
+  /** @type {Array<string>} */
   const imports = freeze([]);
 
   /**